improve dragging, add event to window, not to individual box to prevent lose box
